+++ src/engraving/dom/guitarbend.h
@@ -50,7 +50,6 @@ public:
     LineSegment* createLineSegment(System* parent) override;
 
     Note* startNote() const;
-    void setStartNotePitch(int pitch);
     Note* startNoteOfChain() const;
 
     Note* endNote() const;
