+++ src/inspector/inspectormodule.cpp
@@ -32,8 +32,8 @@
 #include "models/inspectorpopupcontroller.h"
 
 #include "view/widgets/fretcanvas.h"
+#include "view/widgets/bendgridcanvas.h"
 #include "view/widgets/gridcanvas.h"
-#include "view/widgets/compat/gridcanvas.h"
 
 #include "types/directiontypes.h"
 #include "types/slurtietypes.h"
@@ -110,8 +110,8 @@ void InspectorModule::registerUiTypes()
     qmlRegisterUncreatableType<AmbitusTypes>("MuseScore.Inspector", 1, 0, "AmbitusTypes", "Not creatable as it is an enum type");
     qmlRegisterUncreatableType<ChordSymbolTypes>("MuseScore.Inspector", 1, 0, "ChordSymbolTypes", "Not creatable as it is an enum type");
     qmlRegisterType<FretCanvas>("MuseScore.Inspector", 1, 0, "FretCanvas");
+    qmlRegisterType<BendGridCanvas>("MuseScore.Inspector", 1, 0, "BendGridCanvas");
     qmlRegisterType<GridCanvas>("MuseScore.Inspector", 1, 0, "GridCanvas");
-    qmlRegisterType<inspector::compat::GridCanvas>("MuseScore.Inspector", 1, 0, "CompatGridCanvas");
     qmlRegisterUncreatableType<TremoloBarTypes>("MuseScore.Inspector", 1, 0, "TremoloBarTypes", "Not creatable as it is an enum type");
     qmlRegisterUncreatableType<TremoloTypes>("MuseScore.Inspector", 1, 0, "TremoloTypes", "Not creatable as it is an enum type");
     qmlRegisterType<InspectorPopupController>("MuseScore.Inspector", 1, 0, "InspectorPopupController");
