+++ src/inspector/view/widgets/gridcanvas.cpp
@@ -24,7 +24,6 @@
 
 #include <QApplication>
 #include <QPalette>
-#include <QPainterPath>
 #include <cmath>
 
 #include "draw/types/pen.h"
@@ -35,29 +34,15 @@ using namespace mu::inspector;
 using namespace mu::ui;
 using namespace mu::engraving;
 
-static constexpr int GRIP_RADIUS = 6;
-static constexpr int GRIP_CENTER_RADIUS = GRIP_RADIUS - 2;
-static constexpr int GRIP_SELECTED_RADIUS = GRIP_RADIUS + 2;
-static constexpr int GRIP_FOCUS_RADIUS = GRIP_SELECTED_RADIUS + 2;
-
 GridCanvas::GridCanvas(QQuickItem* parent)
     : uicomponents::QuickPaintedView(parent)
 {
     setAcceptedMouseButtons(Qt::AllButtons);
-    setAcceptHoverEvents(true);
-
-    uiConfig()->currentThemeChanged().onNotify(this, [this]() {
-        update();
-    });
-
-    uiConfig()->fontChanged().onNotify(this, [this]() {
-        update();
-    });
 }
 
 QVariant GridCanvas::pointList() const
 {
-    return curvePointsToQVariant(m_points);
+    return pitchValuesToQVariant(m_points);
 }
 
 int GridCanvas::rowCount() const
@@ -137,7 +122,7 @@ void GridCanvas::setShouldShowNegativeRows(bool shouldShowNegativeRows)
 
 void GridCanvas::setPointList(QVariant points)
 {
-    CurvePoints newPointList = curvePointsFromQVariant(points);
+    PitchValues newPointList = pitchValuesFromQVariant(points);
 
     if (m_points == newPointList) {
         return;
@@ -160,412 +145,155 @@ void GridCanvas::paint(QPainter* painter)
              m_columns);
         return;
     }
-
-    QRectF frameRect = this->frameRect();
-
-    drawBackground(painter, frameRect);
-    drawCurve(painter, frameRect);
-}
-
-//---------------------------------------------------------
-//   mousePressEvent
-//---------------------------------------------------------
-
-void GridCanvas::mousePressEvent(QMouseEvent* event)
-{
-    if (!(m_rows && m_columns)) {
-        LOGD("GridCanvas::mousePressEvent: number of columns or rows set to 0.\nColumns: %i, Rows: %i", m_rows,
-             m_columns);
-        return;
-    }
-
-    QRectF frameRect = this->frameRect();
-    std::pair<int, int> coord = this->frameCoord(frameRect, event->pos().x(), event->pos().y());
-    engraving::PitchValue value = pitchValue(frameRect, coord.first, coord.second);
-
-    m_currentPointIndex = this->pointIndex(value);
-}
-
-void GridCanvas::mouseMoveEvent(QMouseEvent* event)
-{
-    if (!m_currentPointIndex.has_value()) {
-        return;
-    }
-
-    QRectF frameRect = this->frameRect();
-    std::pair<int, qreal> coord = this->frameCoord(frameRect, event->pos().x(), event->pos().y());
-
-    engraving::PitchValue value = pitchValue(frameRect, coord.first, coord.second);
-
-    int index = m_currentPointIndex.value();
-
-    bool canMoveHorizontally = m_points[index].canMove(CurvePoint::MoveDirection::Horizontal);
-    bool canMoveVertically = m_points[index].canMove(CurvePoint::MoveDirection::Vertical);
-
-    if (!canMoveHorizontally && !canMoveVertically) {
-        return;
-    }
-
-    if (canMoveVertically) {
-        m_points[index].pitch = value.pitch;
-
-        bool isDashed = m_points[index].endDashed;
-        bool isNextDashed = (index + 1 < m_points.size()) && m_points[index + 1].endDashed;
-
-        if (isDashed) {
-            m_points[index - 1].pitch = value.pitch;
-        }
-
-        if (isNextDashed) {
-            m_points[index + 1].pitch = value.pitch;
-        }
-    }
-
-    if (canMoveHorizontally) {
-        bool canMove = false;
-        bool moveToLeft = m_points[index].time > value.time;
-        if (moveToLeft) {
-            canMove = (index - 1 >= 0) && m_points[index - 1].time < value.time;
-        } else {
-            canMove = (index + 1 < m_points.size()) && m_points[index + 1].time > value.time;
-        }
-
-        if (canMove) {
-            m_points[index].time = value.time;
-        }
-    }
-
-    update();
-    emit canvasChanged();
-}
-
-void GridCanvas::mouseReleaseEvent(QMouseEvent*)
-{
-    m_currentPointIndex = std::nullopt;
-}
-
-void GridCanvas::hoverEnterEvent(QHoverEvent*)
-{
-    m_hoverPointIndex = std::nullopt;
-}
-
-void GridCanvas::hoverMoveEvent(QHoverEvent* event)
-{
-    auto oldPointIndex = m_hoverPointIndex;
-
-    QRectF frameRect = this->frameRect();
-    std::pair<int, int> coord = this->frameCoord(frameRect, event->pos().x(), event->pos().y());
-    engraving::PitchValue value = pitchValue(frameRect, coord.first, coord.second);
-
-    m_hoverPointIndex = this->pointIndex(value);
-
-    if (oldPointIndex != m_hoverPointIndex) {
-        update();
-    }
-}
-
-void GridCanvas::hoverLeaveEvent(QHoverEvent*)
-{
-    m_hoverPointIndex = std::nullopt;
-}
-
-QRectF GridCanvas::frameRect() const
-{
     // not qreal here, even though elsewhere yes,
     // because width and height return a number of pixels,
     // hence integers.
     const int w = width();
     const int h = height();
 
+    const qreal columnWidth = qreal(w) / m_columns;
+    const qreal rowHeight = qreal(h) / m_rows;
+
     // let half a column of margin around
-    const qreal margin = 12.0;
-    const qreal leftPos = margin * 3.0;   // also left margin
-    const qreal topPos = margin;      // also top margin
-    const qreal rightPos = qreal(w) - margin;   // right end position of graph
-    const qreal bottomPos = qreal(h) - margin;   // bottom end position of graph
-
-    return QRectF(QPointF(leftPos, topPos), QPointF(rightPos, bottomPos));
-}
-
-qreal GridCanvas::columnWidth(const QRectF& frameRect) const
-{
-    return frameRect.width() / (m_columns - 1);
-}
-
-qreal GridCanvas::rowHeight(const QRectF& frameRect) const
-{
-    return frameRect.height() / (m_rows - 1);
-}
-
-std::pair<int, int> GridCanvas::frameCoord(const QRectF& frameRect, int x, int y) const
-{
-    // restrict to clickable area
-    if (x > frameRect.right()) {
-        x = frameRect.right();
-    } else if (x < frameRect.left()) {
-        x = frameRect.left();
-    }
-    if (y > frameRect.bottom()) {
-        y = frameRect.bottom();
-    } else if (y < frameRect.top()) {
-        y = frameRect.top();
-    }
-
-    return { x - frameRect.left(), y - frameRect.top() };
-}
-
-void GridCanvas::drawBackground(QPainter* painter, const QRectF& frameRect)
-{
-    const qreal rowHeight = this->rowHeight(frameRect);
-    const qreal columnWidth = this->columnWidth(frameRect);
-
-    const ThemeInfo& currentTheme = uiConfig()->currentTheme();
-    QColor primaryLinesColor(currentTheme.codeKey == DARK_THEME_CODE ? Qt::white : Qt::black);
-    QColor secondaryLinesColor(Qt::gray);
+    const qreal leftPos = columnWidth * .5;   // also left margin
+    const qreal topPos = rowHeight * .5;      // also top margin
+    const qreal rightPos = w - leftPos;   // right end position of graph
+    const qreal bottomPos = h - topPos;   // bottom end position of graph
 
     painter->setRenderHint(QPainter::Antialiasing, true);
 
-    QColor backgroundColor(currentTheme.values[BACKGROUND_PRIMARY_COLOR].toString());
-    painter->fillRect(QRect(0, 0, width(), height()), backgroundColor);
-
+    painter->fillRect(childrenRect(), QApplication::palette().color(QPalette::Window).lighter());
     QPen pen = painter->pen();
     pen.setWidth(1);
 
+    QColor primaryLinesColor(uiConfig()->currentTheme().codeKey == DARK_THEME_CODE ? Qt::white : Qt::black);
+    QColor secondaryLinesColor(Qt::gray);
     // draw vertical lines
-    for (int i = 1; i < m_columns - 1; ++i) {
-        qreal xpos = frameRect.left() + i * columnWidth;
+    for (int i = 0; i < m_columns; ++i) {
+        qreal xpos = leftPos + i * columnWidth;
         // lighter middle lines
         pen.setColor(i % m_primaryColumnsInterval ? secondaryLinesColor : primaryLinesColor);
         painter->setPen(pen);
-        painter->drawLine(xpos, frameRect.top(), xpos, frameRect.bottom());
+        painter->drawLine(xpos, topPos, xpos, bottomPos);
     }
 
     // draw horizontal lines
-    QFont font;
-    font.setFamily(QString::fromStdString(uiConfig()->fontFamily()));
-    font.setPixelSize(uiConfig()->fontSize());
-    int stringHeight = QFontMetrics(font).height();
-    painter->setFont(font);
-
-    int lastPrimaryRowIndex = 0;
-
-    for (int i = 1; i < m_rows - 1; ++i) {
-        int ypos = frameRect.top() + i * rowHeight;
-
-        bool isPrimary = !(i % m_primaryRowsInterval);
-
+    for (int i = 0; i < m_rows; ++i) {
+        int ypos = topPos + i * rowHeight;
         // lighter middle lines
-        pen.setColor(isPrimary ? primaryLinesColor : secondaryLinesColor);
+        pen.setColor(i % m_primaryRowsInterval ? secondaryLinesColor : primaryLinesColor);
         if (m_showNegativeRows) {
             pen.setWidth(i == (m_rows - 1) / 2 ? 3 : 1);
         }
         painter->setPen(pen);
-        painter->drawLine(frameRect.left(), ypos, frameRect.right(), ypos);
-
-        int interval = (m_primaryRowsInterval - 1) - i / m_primaryRowsInterval;
-        bool negative = false;
-
-        if (m_showNegativeRows) {
-            int curveRowMiddleIndex = m_rows / 2;
-            negative = i > curveRowMiddleIndex;
-
-            if (negative) {
-                interval = -(i - curveRowMiddleIndex) / m_primaryRowsInterval;
-            } else {
-                interval = (curveRowMiddleIndex - i) / m_primaryRowsInterval;
-            }
-        }
-
-        bool isHalf = !((i + (i - lastPrimaryRowIndex)) % m_primaryRowsInterval) && !isPrimary;
-
-        if (isPrimary) {
-            lastPrimaryRowIndex = i;
-        }
-
-        if (!isPrimary && !isHalf) {
-            continue;
-        }
-
-        if (!m_showNegativeRows && isHalf) {
-            --interval;
-        }
-
-        pen.setColor(primaryLinesColor);
-        painter->setPen(pen);
-
-        QString intervalStr = QString::number(interval);
-        if (interval == 0 && isHalf) {
-            intervalStr = negative ? "-" : "";
-        }
-
-        QString text = QString("%1%2")
-                       .arg(intervalStr)
-                       .arg(isHalf ? "½" : "");
-
-        QRect textRect(0, ypos - stringHeight / 2, frameRect.left(), stringHeight);
-
-        painter->drawText(textRect, Qt::AlignCenter, text);
+        painter->drawLine(leftPos, ypos, rightPos, ypos);
     }
 
-    // draw a frame
-    QPainterPath path;
-    path.addRoundedRect(frameRect, 3, 3);
-
-    pen.setColor(primaryLinesColor);
-    pen.setWidth(1);
-    pen.setStyle(Qt::PenStyle::SolidLine);
-    painter->setPen(pen);
-
-    painter->fillPath(path, Qt::transparent);
-    painter->drawPath(path);
-}
-
-void GridCanvas::drawCurve(QPainter* painter, const QRectF& frameRect)
-{
-    const qreal rowHeight = this->rowHeight(frameRect);
-    const ThemeInfo& currentTheme = uiConfig()->currentTheme();
-    QColor backgroundColor(currentTheme.values[BACKGROUND_PRIMARY_COLOR].toString());
-
     // this lambda takes as input a pitch value, and determines where what are its x and y coordinates
-    auto getPosition = [this, rowHeight, &frameRect](const CurvePoint& v) -> QPointF {
-        const qreal x = round(qreal(v.time) * (frameRect.width() / PitchValue::MAX_TIME)) + frameRect.left();
+    auto getPosition = [this, columnWidth, rowHeight, leftPos, topPos, bottomPos](const mu::engraving::PitchValue& v) -> QPointF {
+        const qreal x = round((qreal(v.time) / 60) * (m_columns - 1)) * columnWidth + leftPos;
         qreal y = 0;
         if (m_showNegativeRows) {                    // get the middle pos and add the top margin and half of the rows
-            y = frameRect.top() + rowHeight * (m_rows - 1) * .5;
+            y = topPos + rowHeight * (m_rows - 1) * .5;
         } else {                    // from the bottom
-            y = frameRect.bottom();
+            y = bottomPos;
         }
         // add the offset
-        y -=  (qreal(v.pitch) / (100 * (m_rows / m_primaryRowsInterval)) * (m_rows - 1))
+        y -=  round((qreal(v.pitch) / (100 * (m_rows / m_primaryRowsInterval))) * (m_rows - 1))
              * rowHeight;
         return QPointF(x, y);
     };
 
+    static constexpr int GRIP_HALF_RADIUS = 5;
     QPointF lastPoint(0, 0);
-    QPen pen = painter->pen();
+    pen = painter->pen();
     pen.setWidth(3);
-
-    QColor color(currentTheme.values[ACCENT_COLOR].toString());
-    pen.setColor(color);
+    pen.setColor(Qt::red);   // not theme dependant
     painter->setPen(pen);
-
     // draw line between points
-    for (const CurvePoint& v : m_points) {
+    for (const mu::engraving::PitchValue& v : m_points) {
         QPointF currentPoint = getPosition(v);
-
-        QPainterPath path;
-        path.moveTo(lastPoint);
-
         // draw line only if there is a point before the current one
         if (lastPoint.x()) {
-            QPointF c(currentPoint.x(), lastPoint.y());
-            path.quadTo(c, currentPoint);
-
-            if (v.endDashed) {
-                pen.setColor(backgroundColor);
-                painter->strokePath(path, pen);
-            }
-
-            pen.setColor(color);
-            pen.setStyle(v.endDashed ? Qt::PenStyle::DashLine : Qt::PenStyle::SolidLine);
-            painter->strokePath(path, pen);
-
-            pen.setStyle(Qt::PenStyle::SolidLine);
+            painter->drawLine(lastPoint, currentPoint);
         }
-
         lastPoint = currentPoint;
     }
 
-    // draw points
-    QBrush backgroundBrush(backgroundColor, Qt::SolidPattern);
-    QBrush activeBrush(color, Qt::SolidPattern);
-
-    QColor hoverColor(color);
-    hoverColor.setAlpha(200);
-    QBrush hoverBrush(hoverColor, Qt::SolidPattern);
-
     painter->setPen(Qt::NoPen);
-
-    for (int i = 0; i < m_points.size(); ++i) {
-        const CurvePoint& v = m_points[i];
-        if (!v.canMove()) {
-            continue;
-        }
-
-        QPointF pos = getPosition(v);
-
-        bool isNotActiveButton = (!m_hoverPointIndex.has_value() || m_hoverPointIndex.value() != i)
-                                 && (!m_currentPointIndex.has_value() || m_currentPointIndex.value() != i);
-
-        if (isNotActiveButton) { // normal
-            painter->setBrush(activeBrush);
-            painter->drawEllipse(pos, GRIP_RADIUS, GRIP_RADIUS);
-
-            painter->setBrush(backgroundBrush);
-            painter->drawEllipse(pos, GRIP_CENTER_RADIUS, GRIP_CENTER_RADIUS);
-        } else if (m_hoverPointIndex.has_value() && m_currentPointIndex.has_value()) { // focused
-            QColor fontPrimaryColor(currentTheme.values[FONT_PRIMARY_COLOR].toString());
-            QBrush fontPrimaryBrush(fontPrimaryColor, Qt::SolidPattern);
-            painter->setBrush(fontPrimaryBrush);
-            painter->drawEllipse(pos, GRIP_FOCUS_RADIUS, GRIP_FOCUS_RADIUS);
-
-            painter->setBrush(backgroundBrush);
-            painter->drawEllipse(pos, GRIP_SELECTED_RADIUS, GRIP_SELECTED_RADIUS);
-
-            painter->setBrush(activeBrush);
-            painter->drawEllipse(pos, GRIP_RADIUS, GRIP_RADIUS);
-        } else if (m_hoverPointIndex.has_value() && m_hoverPointIndex.value() == i) { // hover
-            painter->setBrush(activeBrush);
-            painter->drawEllipse(pos, GRIP_RADIUS, GRIP_RADIUS);
-
-            painter->setBrush(hoverBrush);
-            painter->drawEllipse(pos, GRIP_CENTER_RADIUS, GRIP_CENTER_RADIUS);
-        } else if (m_currentPointIndex.has_value() && m_currentPointIndex.value() == i) { // selected
-            painter->setBrush(backgroundBrush);
-            painter->drawEllipse(pos, GRIP_SELECTED_RADIUS, GRIP_SELECTED_RADIUS);
-
-            painter->setBrush(activeBrush);
-            painter->drawEllipse(pos, GRIP_RADIUS, GRIP_RADIUS);
-        }
+    painter->setBrush(QColor::fromRgb(32, 116, 189));   // Musescore blue
+    // draw points
+    for (const mu::engraving::PitchValue& v : m_points) {
+        painter->drawEllipse(getPosition(v), GRIP_HALF_RADIUS, GRIP_HALF_RADIUS);
     }
 }
 
-std::optional<int> GridCanvas::pointIndex(const engraving::PitchValue& pitchValue, bool movable) const
+//---------------------------------------------------------
+//   mousePressEvent
+//---------------------------------------------------------
+
+void GridCanvas::mousePressEvent(QMouseEvent* ev)
 {
-    const int numberOfPoints = m_points.size();
+    if (!(m_rows && m_columns)) {
+        LOGD("GridCanvas::mousePressEvent: number of columns or rows set to 0.\nColumns: %i, Rows: %i", m_rows,
+             m_columns);
+        return;
+    }
+    const qreal columnWidth = qreal(width()) / m_columns;
+    const qreal rowHeight = qreal(height()) / m_rows;
 
-    for (int i = 0; i < numberOfPoints; ++i) {
-        const CurvePoint& point = m_points[i];
-        if (movable != point.canMove()) {
-            continue;
-        }
+    // Half a column/row of margin around
+    const int x = ev->x() - columnWidth * .5;
+    const int y = ev->y() - rowHeight * .5;
 
-        if (std::pow((pitchValue.time - point.time), 2)
-            + std::pow((pitchValue.pitch - point.pitch), 2) < std::pow(GRIP_CENTER_RADIUS, 2)) {
-            return i;
-        }
+    int column = round(qreal(x) / columnWidth);
+    int row = round(qreal(y) / rowHeight);
+
+    // restrict to clickable area
+    if (column >= m_columns) {
+        column = m_columns - 1;
+    } else if (column < 0) {
+        column = 0;
+    }
+    if (row >= m_rows) {
+        row = m_rows - 1;
+    } else if (row < 0) {
+        row = 0;
     }
 
-    return std::nullopt;
-}
-
-PitchValue GridCanvas::pitchValue(const QRectF& frameRect, int frameX, int frameY) const
-{
-    PitchValue value;
-    value.time = qreal(frameX) / (frameRect.width() / PitchValue::MAX_TIME);
-
-    const qreal rowHeight = this->rowHeight(frameRect);
-    int row = m_rows - 1 - round(qreal(frameY) / rowHeight);
+    // invert y.
     if (m_showNegativeRows) {
-        int half = (m_rows - 1) / 2;
-        if (row > half) {
-            row -= half;
-        } else {
-            row += -half;
-        }
+        row = (m_rows - 1) / 2 - row;
+    } else {
+        row = (m_rows - 1) - row;
     }
 
-    value.pitch = row * 100 / m_primaryRowsInterval;
+    const int time = column * 60 / (m_columns - 1);
+    const int pitch = row * 100 / m_primaryRowsInterval;
 
-    return value;
+    const size_t numberOfPoints = m_points.size();
+    bool found = false;
+    for (size_t i = 0; i < numberOfPoints; ++i) {
+        if (round(qreal(m_points[i].time) / 60 * (m_columns - 1)) > column) {
+            m_points.insert(m_points.begin() + i, mu::engraving::PitchValue(time, pitch, false));
+            found = true;
+            break;
+        }
+        if (round(qreal(m_points[i].time) / 60 * (m_columns - 1)) == column) {
+            if (round(qreal(m_points[i].pitch) / (100 * (m_rows / m_primaryRowsInterval)) * (m_rows - 1)) == row
+                && i > 0 && i < (numberOfPoints - 1)) {
+                m_points.erase(m_points.begin() + i);
+            } else {
+                m_points[i].pitch = pitch;
+            }
+            found = true;
+            break;
+        }
+    }
+    if (!found) {
+        m_points.push_back(mu::engraving::PitchValue(time, pitch, false));
+    }
+
+    update();
+    emit canvasChanged();
 }
